Remove `no-confusing-arrow` because of the deprecation of formatting rules
First, this rule was deprecated in ESLint v8.53.0.
The alternative is the third-party plugin `@stylistic/eslint-plugin-js`.

Second, I sometimes feel annoying on this rule. For example, see:

```js
const result = () => (type === 'foo' ? foo() : bar());
/*             ^^ Arrow function used ambiguously with a conditional expression [no-confusing-arrow] */
```

In this case, I think it's enough readable, thanks to `()` around the function body.
So, I think this rule is no longer necessary in this shared config.

Ref:
- https://eslint.org/docs/latest/rules/no-confusing-arrow
- https://eslint.org/blog/2023/10/deprecating-formatting-rules
- https://eslint.style/packages/js